Move `std::io::cursor`internals to `core::io`
