+++ library/std/src/io/cursor.rs
@@ -4,8 +4,11 @@ mod tests;
 #[stable(feature = "rust1", since = "1.0.0")]
 pub use core::io::Cursor;
 
+use alloc_crate::io::{
+    slice_write, slice_write_all, slice_write_all_vectored, slice_write_vectored,
+};
+
 use crate::alloc::Allocator;
-use crate::cmp;
 use crate::io::prelude::*;
 use crate::io::{self, BorrowedCursor, ErrorKind, IoSlice, IoSliceMut};
 
@@ -148,53 +151,6 @@ impl<W: WriteThroughCursor> Write for Cursor<W> {
     fn flush(&mut self) -> io::Result<()> {
         WriteThroughCursor::flush(self)
     }
-}
-
-// Non-resizing write implementation
-#[inline]
-fn slice_write(pos_mut: &mut u64, slice: &mut [u8], buf: &[u8]) -> io::Result<usize> {
-    let pos = cmp::min(*pos_mut, slice.len() as u64);
-    let amt = (&mut slice[(pos as usize)..]).write(buf)?;
-    *pos_mut += amt as u64;
-    Ok(amt)
-}
-
-#[inline]
-fn slice_write_vectored(
-    pos_mut: &mut u64,
-    slice: &mut [u8],
-    bufs: &[IoSlice<'_>],
-) -> io::Result<usize> {
-    let mut nwritten = 0;
-    for buf in bufs {
-        let n = slice_write(pos_mut, slice, buf)?;
-        nwritten += n;
-        if n < buf.len() {
-            break;
-        }
-    }
-    Ok(nwritten)
-}
-
-#[inline]
-fn slice_write_all(pos_mut: &mut u64, slice: &mut [u8], buf: &[u8]) -> io::Result<()> {
-    let n = slice_write(pos_mut, slice, buf)?;
-    if n < buf.len() { Err(io::Error::WRITE_ALL_EOF) } else { Ok(()) }
-}
-
-#[inline]
-fn slice_write_all_vectored(
-    pos_mut: &mut u64,
-    slice: &mut [u8],
-    bufs: &[IoSlice<'_>],
-) -> io::Result<()> {
-    for buf in bufs {
-        let n = slice_write(pos_mut, slice, buf)?;
-        if n < buf.len() {
-            return Err(io::Error::WRITE_ALL_EOF);
-        }
-    }
-    Ok(())
 }
 
 /// Reserves the required space, and pads the vec with 0s if necessary.
